Move WF optimization driver code to WFOpt
