+++ src/QMCDrivers/WFOpt/QMCCostFunctionBase.h
