+++ src/QMCDrivers/WFOpt/QMCLinearOptimize.h
